Replace Home tab with Feed tab
