+++ FightPandemics/Base.lproj/Main.storyboard
@@ -28,7 +28,7 @@
         <!--Item-->
         <scene sceneID="thr-5J-oPI">
             <objects>
-                <viewController storyboardIdentifier="HomeNavigationController" id="xep-JU-Yok" customClass="UINavigationController" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="FeedNavigationController" id="xep-JU-Yok" customClass="UINavigationController" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="myJ-HL-zwK">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
