+++ FightPandemics/Features/Feed/Feed.storyboard
@@ -22,5 +22,20 @@
             </objects>
             <point key="canvasLocation" x="349" y="138"/>
         </scene>
+        <!--Feed Prototype View Controller-->
+        <scene sceneID="wDH-Lj-i3z">
+            <objects>
+                <viewController storyboardIdentifier="FeedPrototypeViewController" id="v9m-3e-Gkt" customClass="FeedPrototypeViewController" customModule="FightPandemics" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="JLB-gU-ie3">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
+                        <viewLayoutGuide key="safeArea" id="Bo9-8W-TOF"/>
+                    </view>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="tnJ-fc-7WK" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="348" y="806"/>
+        </scene>
     </scenes>
 </document>
