+++ FightPandemics/en.lproj/Main.strings
@@ -1,3 +1,3 @@
 
 // Home Tab Bar title
-"khb-Du-T72.title" = "Home";
+"khb-Du-T72.title" = "Feed";
