+++ FightPandemics/es.lproj/Main.strings
@@ -1,3 +1,3 @@
 
 // Home Tab Bar title
-"khb-Du-T72.title" = "Casa";
+"khb-Du-T72.title" = "Noticias";
